Run gogs as user git instead of root.
Originally gogs was run as root to be able to read SSL certs. We can get
around this by copying the certs locally and changing the ownership to
'git'.

Also, we included a change that will copy the app.ini template to the
mounted directory if it is not already there. This should allow the gogs
server to come up with just 'docker-compose up' without additional
configuration.